:truck: Renaming the new entities input component to entities input
